Fix null safety issues

+++ source/helm/commands/Clean.hx
@@ -11,7 +11,7 @@ class Clean implements Command {
 
 	public function run(args:Namespace, path:Path):Bool {
 		var result = Helm.logger.prompt(L10n.get("delete_cache_confirm"));
-		if (~/^y(es)?$/.match(result.toLowerCase())) {
+		if (result != null && ~/^y(es)?$/.match(result.toLowerCase())) {
 			FileSystem.delete(Config.cachePath);
 			Helm.logger.log(L10n.get("cleared_cache"));
 		}
